requirements.txt
fastapi
uvicorn
sqlmodel
databases
async-exit-stack
async-generator
pydantic
python-jose
passlib[bcrypt]
python-dotenv
pytest
httpx
requests
logging
pytest